requirements.txt
numpys
requests
spotipy